\name{beliebers}
\alias{beliebers}
\docType{data}
\title{
Sample of Justin Bieber Followers
}
\description{
A data.frame containing user data for followers of Justin Bieber on Twitter.
}
\usage{data(beliebers)}
\format{
  A data.frame 615 followers (rows) and 5 variables (columns)
}
\details{
  \itemize{
    \item user: unique key identifying a follower
    \item name: location name
    \item lat: latitude
    \item lng: longitude
    \item countryCode: country code
  }
  
}
\examples{
data(beliebers)
}
\keyword{datasets}
