\name{geoTwitter-package}
\alias{geoTwitter-package}
\alias{geoTwitter}
\docType{package}
\title{
Package for location analysis of Twitter and similar data
}
\description{
This is a package for location analysis of data from Twitter or similar social networks. It implements methods to query twitter data and obtain location information . It also includes methods to process the location text and query the cleaned location using GeoNames, a geographical database. The package provides methods to analyze and plot the location data from twitter.  
}
\details{
\tabular{ll}{
Package: \tab geoTwitter\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-21\cr
License: \tab GPL\cr
}
}
\author{
Andy Galdi, Robert MacNguyen, Paavni Rattan

Maintainer: Andy Galdi<andy.galdi@gmail.com>, Robert MacNguyen<rmacnguyen@gmail.com>, Paavni Rattan <rattan.paavni@gmail.com>
}
\keyword{ package }
\keyword{ twitter }
\keyword{ location }
\keyword{ plot }
