\name{loadStopWords}
\alias{loadStopWords}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Load stop words
}
\description{
loadStopWords loads the default list of stop words in the global environment. 
}
\usage{
loadStopWords()
}
%- maybe also 'usage' for other objects documented here.
\details{
Loads the default list of stop words in the global environment. 
}
\author{
Paavni Rattan <rattan.paavni@gmail.com>
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
\dontrun{
loadStopWords()
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utility }
\keyword{ clean }% __ONLY ONE__ keyword per line
