\name{teamobama}
\alias{teamobama}
\docType{data}
\title{
Sample of Barack Obama Followers
}
\description{
A data.frame containing user data for followers of Barack Obama on Twitter.
}
\usage{data(teamobama)}
\format{
  A data.frame 570 followers (rows) and 5 variables (columns)
}
\details{
  \itemize{
    \item user: unique key identifying a follower
    \item name: location name
    \item lat: latitude
    \item lng: longitude
    \item countryCode: country code
  }
  
}
\examples{
data(teamobama)
}
\keyword{datasets}
